requirements.txt
fabric
pyyaml
